Add challenge quests list in challenge activity. Cleanup.
