+++ app/src/main/java/io/ipoli/android/app/AppComponent.java
@@ -30,8 +30,6 @@ import io.ipoli.android.challenge.activities.PickChallengeActivity;
 import io.ipoli.android.challenge.activities.PickChallengeQuestsActivity;
 import io.ipoli.android.challenge.activities.PickDailyChallengeQuestsActivity;
 import io.ipoli.android.challenge.fragments.ChallengeListFragment;
-import io.ipoli.android.challenge.fragments.ChallengeOverviewFragment;
-import io.ipoli.android.challenge.fragments.ChallengeQuestListFragment;
 import io.ipoli.android.challenge.receivers.DailyChallengeReminderReceiver;
 import io.ipoli.android.challenge.receivers.ScheduleDailyChallengeReminderReceiver;
 import io.ipoli.android.pet.PetActivity;
@@ -149,10 +147,6 @@ public interface AppComponent {
     void inject(RepeatingQuestActivity repeatingQuestActivity);
 
     void inject(ChallengeActivity challengeActivity);
-
-    void inject(ChallengeOverviewFragment challengeOverviewFragment);
-
-    void inject(ChallengeQuestListFragment challengeQuestListFragment);
 
     void inject(PickChallengeQuestsActivity pickChallengeQuestsActivity);
