--- app/src/main/res/layout/fragment_challenge_quest_list.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/root_container"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/md_white">
-
-    <io.ipoli.android.app.ui.EmptyStateRecyclerView
-        android:id="@+id/quest_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/md_white"
-        android:scrollbars="vertical" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/add_quests"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="16dp"
-        android:src="@drawable/ic_add_white_24dp" />
-
-</android.support.design.widget.CoordinatorLayout>
