+++ app/src/main/res/values/strings.xml
@@ -343,4 +343,5 @@
     <string name="challenge_accepted">Challenge accepted!</string>
     <string name="loading">Loading...</string>
     <string name="widget_agenda_empty">iPoli wants a quest!</string>
+    <string name="quests">Quests</string>
 </resources>
